requirements.txt
slack_bolt>=1.19.0
python-dotenv>=1.0.0
openai>=1.0.0
supabase>=2.0.0
slack-sdk
flask
anthropic>=0.40.0